add function addSong
add help function :
1) isIdExist - Does an ID exist in the songs
2) maxID - find max ID in song

function addSong
Gets a title, album, artist, duration & ID. Adds a new song with given properties to the player. The ID is optional, and if omitted should be automatically generated. The song duration should be in `mm:ss` format (for example 06:27). Returns the ID of the new song.