add createPlaylist function
call help function
isIdExist- Does an ID exist in the playlist
maxID- find max ID in playlist

the createPlaylist
Gets a name & ID. Creates a new, empty playlist with the given details. The ID is optional, and if omitted should be automatically generated. Returns the ID of the new playlist.