add editPlaylist function
first i throw if there isn't song/playlist ID.
than i push If the song ID does not exist in the playlist, remove if the song ID exists in the playlist, or remove the playlist if the song is only in the playlist.

Gets a playlist ID & a song ID. If the song ID exists in the playlist, removes it. If it was the only song in the playlist, also deletes the playlist. If the song ID does not exist in the playlist, adds it to the end of the playlist.